Add with sorting Hoc
- Add cities-places component
- Compose sorting and cards list components
- Brings cities-places sorting type state in Hoc
